api/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
python-multipart
filelock
websockets
